refactor(db): colocate IndexedDB schemas with web storage layer
Move RecordingsDbSchemaV1-V5, SerializedAudio, and
RecordingStoredInIndexedDB from models/recordings.ts into
web/dexie-schemas.ts—these types are consumed exclusively by the
web/ Dexie code and were adding 72 lines of noise to what should
be a clean domain type file.

recordings.ts now contains only the Recording domain type (34 lines).
The models barrel no longer re-exports IndexedDB-specific types.
