+++ apps/whispering/src/lib/services/db/models/recordings.ts
@@ -5,7 +5,7 @@
  * It is NOT the storage format - different storage implementations use different formats:
  *
  * - Desktop: Stores metadata in markdown files (.md) and audio in separate files (.webm, .mp3)
- * - Web: Stores in IndexedDB with serialized audio (see RecordingStoredInIndexedDB)
+ * - Web: Stores in IndexedDB with serialized audio (see web/dexie-schemas.ts)
  *
  * Both implementations read their storage format and convert it to this intermediate
  * representation for use in the UI layer.
@@ -31,76 +31,4 @@ export type Recording = {
 	 * 4. Marked as 'FAILED' if transcription fails
 	 */
 	transcriptionStatus: 'UNPROCESSED' | 'TRANSCRIBING' | 'DONE' | 'FAILED';
-};
-
-/**
- * Serialized audio format for IndexedDB storage.
- *
- * This format is used to work around iOS Safari's limitations with storing Blob objects
- * in IndexedDB. Instead of storing the Blob directly (which can fail or become corrupted
- * on iOS), we deconstruct it into:
- * - arrayBuffer: The raw binary data
- * - blobType: The original MIME type (e.g., 'audio/webm', 'audio/wav')
- *
- * This can be reliably stored in IndexedDB on all platforms, including iOS Safari.
- * To reconstruct: new Blob([arrayBuffer], { type: blobType })
- *
- * @see /Users/braden/Code/whispering/.conductor/la-paz/docs/patterns/serialized-audio-pattern.md
- */
-export type SerializedAudio = {
-	arrayBuffer: ArrayBuffer;
-	blobType: string;
-};
-
-/**
- * How a recording is actually stored in IndexedDB (storage format).
- *
- * This is NOT the intermediate representation used by the UI (see Recording type).
- *
- * Extends Recording with:
- * - `serializedAudio` field for storing audio data (see SerializedAudio type)
- * - Audio is stored in serialized format to work around iOS Safari Blob storage issues
- */
-export type RecordingStoredInIndexedDB = Recording & {
-	serializedAudio: SerializedAudio | undefined;
-};
-
-export type RecordingsDbSchemaV5 = {
-	recordings: RecordingStoredInIndexedDB;
-};
-
-export type RecordingsDbSchemaV4 = {
-	recordings: RecordingsDbSchemaV3['recordings'] & {
-		// V4 added 'createdAt' and 'updatedAt' fields
-		createdAt: string;
-		updatedAt: string;
-	};
-};
-
-export type RecordingsDbSchemaV3 = {
-	recordings: RecordingsDbSchemaV1['recordings'];
-};
-
-export type RecordingsDbSchemaV2 = {
-	recordingMetadata: Omit<RecordingsDbSchemaV1['recordings'], 'blob'>;
-	recordingBlobs: { id: string; blob: Blob | undefined };
-};
-
-export type RecordingsDbSchemaV1 = {
-	recordings: {
-		id: string;
-		title: string;
-		subtitle: string;
-		timestamp: string;
-		transcribedText: string;
-		blob: Blob | undefined;
-		/**
-		 * A recording
-		 * 1. Begins in an 'UNPROCESSED' state
-		 * 2. Moves to 'TRANSCRIBING' while the audio is being transcribed
-		 * 3. Finally is marked as 'DONE' when the transcription is complete.
-		 * 4. If the transcription fails, it is marked as 'FAILED'
-		 */
-		transcriptionStatus: 'UNPROCESSED' | 'TRANSCRIBING' | 'DONE' | 'FAILED';
-	};
 };
